requirements.txt
# Django Core
Django==4.2.24
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# Database (PostgreSQL for Railway)
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Production Server
gunicorn==21.2.0
whitenoise==6.10.0

# Other essential packages
Pillow==11.3.0
python-dotenv==1.0.1
requests==2.32.5